% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{addArgs}
\alias{addArgs}
\title{Function to add the correct args onto a polynomial, dummy, or lspline term
Used to live in pasting_args.R}
\usage{
addArgs(varname, var, type)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Function to add the correct args onto a polynomial, dummy, or lspline term
Used to live in pasting_args.R
}
\keyword{internal}
