% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{checkNesting}
\alias{checkNesting}
\title{Used to live in pasting_args.R}
\usage{
checkNesting(col, nested)
}
\arguments{
\item{nested}{}
}
\value{

}
\description{
Used to live in pasting_args.R
}
\keyword{internal}
