% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{createCols}
\alias{createCols}
\title{Function to create the correct columns for processTerm
Takes a predictor and the terms vector, determines which it is,
if an interaction, checks both
don't put in the intercept
Used to live in pasting_args.R}
\usage{
createCols(pds, tms)
}
\arguments{
\item{pds}{predictor}

\item{tms}{terms vector}
}
\value{

}
\description{
Function to create the correct columns for processTerm
Takes a predictor and the terms vector, determines which it is,
if an interaction, checks both
don't put in the intercept
Used to live in pasting_args.R
}
\keyword{internal}
