% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{describe_stratified_date}
\alias{describe_stratified_date}
\title{Describe a Date object within strata}
\usage{
describe_stratified_date(
  x,
  strata,
  subset,
  probs = c(0.25, 0.5, 0.75),
  thresholds = NULL,
  geomInclude = FALSE,
  replaceZeroes = FALSE
)
}
\arguments{
\item{x}{the vector of data to describe}

\item{strata}{the strata}

\item{subset}{a vector indicating a subset within which descriptive statistics
are desired.}

\item{probs}{a vector of proportions to use for quantiles}

\item{thresholds}{a vector of thresholds used to dichotomize variables.}

\item{geomInclude}{if not \code{FALSE} (the default), includes the geometric mean 
in the descriptive statistics.}

\item{replaceZeroes}{if not \code{FALSE} (the default), this indicates a value 
to be used in place of zeroes when computing a geometric mean. If \code{TRUE}, 
a value equal to one-half the lowest nonzero value is used. If a numeric value 
is supplied, that value is used for all variables.}
}
\value{
descriptive statistics within strata for a Date object
}
\description{
Describe a Date object within strata
}
\keyword{internal}
