% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{describe_surv}
\alias{describe_surv}
\title{Describe a Survival variable}
\usage{
describe_surv(
  x,
  probs = c(0.25, 0.5, 0.75),
  thresholds = NULL,
  geometricMean = FALSE,
  geomInclude = FALSE,
  replaceZeroes = FALSE,
  restriction = Inf
)
}
\arguments{
\item{x}{the variable of interest}

\item{probs}{a vector of proportions to use in describing (e.g., 0.25, 0.5)}

\item{thresholds}{a vector of thresholds used to dichotomize variables.}

\item{geometricMean}{logical, should we compute the geomtric mean? Defaults
to \code{FALSE}.}

\item{geomInclude}{if not \code{FALSE}, includes the geometric mean in the 
descriptive statistics. Default is \code{FALSE}.}

\item{replaceZeroes}{if not \code{FALSE}, this indicates a value to be used 
in place of zeroes when computing a geometric mean. If \code{TRUE}, a value 
equal to one-half the lowest nonzero value is used. If a numeric value is 
supplied, that value is used for all variables.}

\item{resctriction}{a value used for computing restricted means, 
standard deviations, and geometric means with censored time to event data. 
The default value of \code{Inf} will cause restrictions at the highest 
observation. Note that the same value is used for all variables of class \code{Surv}.}
}
\value{
A matrix with descriptive statistics for the Surv object
}
\description{
Describe a Survival variable
}
\keyword{internal}
