% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{describe_vector}
\alias{describe_vector}
\title{Descriptive Statistics for a Vector of Data}
\usage{
describe_vector(
  x,
  probs = c(0.25, 0.5, 0.75),
  thresholds = NULL,
  geometricMean = FALSE,
  geomInclude = FALSE,
  replaceZeroes = FALSE
)
}
\arguments{
\item{x}{a vector of data}

\item{probs}{a vector of proportions to use in describing (e.g., 0.25, 0.5)}

\item{thresholds}{a vector of thresholds used to dichotomize variables.}

\item{geometricMean}{logical, should we compute the geomtric mean? Defaults
to \code{FALSE}.}

\item{geomInclude}{if not \code{FALSE}, includes the geometric mean in the 
descriptive statistics. Default is \code{FALSE}.}

\item{replaceZeroes}{if not \code{FALSE}, this indicates a value to be used 
in place of zeroes when computing a geometric mean. If \code{TRUE}, a value 
equal to one-half the lowest nonzero value is used. If a numeric value is 
supplied, that value is used for all variables.}
}
\value{
A matrix of descriptive statistics
}
\description{
Descriptive Statistics for a Vector of Data
}
\keyword{internal}
