% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_diagnostics.R
\name{dfbeta.uRegress}
\alias{dfbeta.uRegress}
\alias{dfbeta}
\title{Calculate dfbeta from \code{uRegress} objects}
\usage{
\method{dfbeta}{uRegress}(model, ...)
}
\arguments{
\item{model}{an object of class \code{uRegress}, as returned by
\link[rigr]{regress}.}

\item{...}{other arguments to pass to \code{stats::dfbeta}}
}
\value{
a matrix of dfbeta values, with a row for each observation and a column for each model coefficient
}
\description{
Extracts dfbeta from \code{uRegress} objects by relying on
functionality from the \code{stats} package. Note that 
\code{dfbeta} and \code{dfbetas} are not the same (\code{dfbetas} are 
less than the \code{dfbeta} values by a 
scaling factor that reflects both the leverage of the observation in
 question and the residual model error).
}
