% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_diagnostics.R
\name{dfbetas.uRegress}
\alias{dfbetas.uRegress}
\title{Calculate dfbetas from \code{uRegress} objects}
\usage{
\method{dfbetas}{uRegress}(model, ...)
}
\arguments{
\item{model}{an object of class \code{uRegress}, as returned by
\link[rigr]{regress}.}

\item{...}{other arguments to pass to \code{stats::dfbetas}}
}
\value{
a matrix of dfbetas values, with a row for each observation and a column for each model coefficient
}
\description{
Extracts dfbetas from \code{uRegress} objects by relying on
functionality from the \code{stats} package. Note that 
\code{dfbeta} and \code{dfbetas} are not the same (\code{dfbetas} are 
less than the \code{dfbeta} values by a 
scaling factor that reflects both the leverage of the observation in
 question and the residual model error).
}
