% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{equal}
\alias{equal}
\title{A function to check if all values in a vector are equal}
\usage{
equal(x)
}
\arguments{
\item{x}{a vector}
}
\value{
\code{TRUE}/\code{FALSE}
}
\description{
A function to check if all values in a vector are equal
}
\keyword{internal}
