% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{explode}
\alias{explode}
\title{Function which takes in a string and returns a vector of individual characters
Helper function for print.uRegress}
\usage{
explode(str)
}
\arguments{
\item{str}{a string}
}
\value{
a vector of individual characters
}
\description{
Function which takes in a string and returns a vector of individual characters
Helper function for print.uRegress
}
\keyword{internal}
