% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fev}
\alias{fev}
\title{FEV dataset}
\format{
A data frame with 654 rows and 7 variables.
}
\source{
\url{http://www.emersonstatistics.com/datasets/fev.txt}
}
\usage{
fev
}
\description{
Data from a study of 654 children on the relationship between smoking status and 
lung function (measured by FEV). Each row corresponds to a single clinic visit and contains
information on age, height, sex, FEV, and smoking status.
More information, including a coding key, is available at
\url{http://www.emersonstatistics.com/datasets/fev.doc}.
}
\keyword{datasets}
