% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{getTerms}
\alias{getTerms}
\title{Takes a formula, returns a list of terms}
\usage{
getTerms(form, y)
}
\arguments{
\item{form}{the formula}

\item{y}{}
}
\value{
the list of terms
}
\description{
Takes a formula, returns a list of terms
}
\keyword{internal}
