% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{insertVec}
\alias{insertVec}
\title{Inserts a vector into another vector}
\usage{
insertVec(x1, x2, indx)
}
\arguments{
\item{indx}{}
}
\value{

}
\description{
Inserts a vector into another vector
}
\keyword{internal}
