% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{is.Date}
\alias{is.Date}
\title{Check if an object is a "Date" object}
\usage{
is.Date(x)
}
\arguments{
\item{x}{the variable}
}
\value{
A logical flag indicating whether the variable is a "Date" variable.
}
\description{
Check if an object is a "Date" object
}
\keyword{internal}
