% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{KM}
\alias{KM}
\title{Compute Kaplan-Meier estimates of the survival curve}
\usage{
KM(x)
}
\arguments{
\item{x}{the data}
}
\value{
Kaplan-Meier estimates of the survival curve
}
\description{
Compute Kaplan-Meier estimates of the survival curve
}
\keyword{internal}
