% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincom.R
\name{lincom}
\alias{lincom}
\alias{lincom.do}
\alias{print.lincom}
\title{Tests of Linear Combinations of Regression
Coefficients}
\usage{
lincom(
  reg,
  comb,
  hyp = 0,
  conf.level = 0.95,
  robustSE = TRUE,
  eform = reg$fnctl != "mean"
)
}
\arguments{
\item{reg}{an object of class \code{uRegress}.}

\item{comb}{a vector or matrix containing the values of the constants which
create the linear combination of the form \deqn{c_0 + c_1\beta_1 + \dots}.
need to give zeroes if coefficients aren't going to be included}

\item{hyp}{the null hypothesis to compare the linear combination of
coefficients against. The default value is \code{0}. An error will be thrown
if the number of columns of this matrix are not equal to the number of
coefficients in the model. a matrix (or single) hypothesis}

\item{conf.level}{a number between 0 and 1, indicating the desired
confidence level for intervals.}

\item{robustSE}{a logical value indicating whether or not to use robust
standard errors in calculation. If \code{TRUE}, then \code{robustSE} must
have been \code{TRUE} when \code{reg} was created.}

\item{eform}{a logical value indicating whether or not to exponentiate the
estimated coefficient. By default this is performed based on the type of
regression used.}
}
\value{
Prints a matrix with the point
estimate of the linear combination of coefficients, a p-value, and
confidence interval.
}
\description{
Produces point estimates, interval estimates, and p values for linear
combinations of regression coefficients using a \code{ uRegress} object.
}
\examples{

# Loading required libraries
library(survival)
library(sandwich)

# Reading in a dataset
data(mri)
attach(mri)

# Linear regression of LDL on age (with robust SE by default)
testReg <- regress ("mean", ldl~age+stroke)

# Testing coefficient created by .5*age - stroke (the first 1 comes from including the intercept)
testC <- c(1, .5, -1)
lincom(testReg, testC)

}
\author{
Scott S. Emerson, M.D., Ph.D., Andrew J. Spieker,
Brian D. Williamson
}
\keyword{~kwd1}
\keyword{~kwd2}
