% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lspline.R
\name{lspline}
\alias{lspline}
\alias{lsplineD}
\title{Create Linear Splines}
\usage{
lspline(x, knots, lbl = NULL, parameterization = "absolute", version = FALSE)
}
\arguments{
\item{x}{\code{y} variable used to create the linear
splines.}

\item{knots}{\code{cluster} vector of knots to create
the splines.}

\item{lbl}{a label for the splines.}

\item{parameterization}{defaults to\code{"absolute"}, and provides splines
based on the absolute slope between knots. If \code{"change"}, provides
splines based on the change from knot to knot. If \code{lsplineD} is called,
\code{"change"} is entered by default.}

\item{version}{if \code{TRUE}, returns the version of the function and
nothing else.}
}
\value{
A matrix containing the linear
splines.
}
\description{
Creates linear splines, mostly for use in regression.
}
\examples{

# Reading in a dataset
data(mri)
attach(mri)
# Create a spline based on absolute
lspline(ldl, c(70, 100, 130, 160))

# Create a spline based on change
lsplineD(ldl, c(70, 100, 130, 160))

}
