% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{mean_KM}
\alias{mean_KM}
\title{Compute the restricted mean survival time}
\usage{
mean_KM(x, restriction = Inf)
}
\arguments{
\item{x}{an object of class \code{KM} (from a call to \code{KM})}

\item{restriction}{a value used for computing restricted means, standard deviations, 
and geometric means with censored time-to-event data. The default value of 
\code{Inf} will cause restrictions at the highest observation. Note that the 
same value is used for all variables of class \code{Surv}.}
}
\value{
the restricted mean survival time
}
\description{
Compute the restricted mean survival time
}
\keyword{internal}
