% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{movingSum}
\alias{movingSum}
\title{Function that sums across a logical vector, returns the current value at each point
Used to live in pasting_args.R}
\usage{
movingSum(vec)
}
\arguments{
\item{vec}{}
}
\value{

}
\description{
Function that sums across a logical vector, returns the current value at each point
Used to live in pasting_args.R
}
\keyword{internal}
