% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{parseFormula}
\alias{parseFormula}
\title{Takes a formula, returns a viable formula}
\usage{
parseFormula(form, modelframe, mat)
}
\arguments{
\item{form}{the formula}

\item{mat}{}
}
\value{
the list
}
\description{
Takes a formula, returns a viable formula
}
\keyword{internal}
