% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{parseList}
\alias{parseList}
\title{Takes in a linearized list, returns a list with only the non-null components}
\usage{
parseList(lst)
}
\arguments{
\item{lst}{}
}
\value{
a list with only the non-null components
}
\description{
Takes in a linearized list, returns a list with only the non-null components
}
\keyword{internal}
