% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{parseParseFormula}
\alias{parseParseFormula}
\title{Go through a parseFactor and get the relevant formulas and partials}
\usage{
parseParseFormula(lst)
}
\arguments{
\item{lst}{a list returned by parseFactor}
}
\value{
a list with two components - the formula and a list of all of the matrices
}
\description{
Go through a parseFactor and get the relevant formulas and partials
}
\keyword{internal}
