% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{parsePartials}
\alias{parsePartials}
\title{Takes a formula, returns a list of models for multiple partial tests}
\usage{
parsePartials(form, modelframe, mat)
}
\arguments{
\item{form}{the formula}

\item{mat}{}
}
\value{
the list
}
\description{
Takes a formula, returns a list of models for multiple partial tests
}
\keyword{internal}
