% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{pasteOn}
\alias{pasteOn}
\title{Function to paste on degree
Used to live in pasting_args.R}
\usage{
pasteOn(x, str, num)
}
\arguments{
\item{num}{}
}
\value{

}
\description{
Function to paste on degree
Used to live in pasting_args.R
}
\keyword{internal}
