% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{pasteOnSpline}
\alias{pasteOnSpline}
\title{Helper function for lspline
Used to live in pasting_args.R}
\usage{
pasteOnSpline(x, nums)
}
\arguments{
\item{nums}{}
}
\value{

}
\description{
Helper function for lspline
Used to live in pasting_args.R
}
\keyword{internal}
