% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{pastePair}
\alias{pastePair}
\title{Function that takes a pair and pastes together, with colon
Used to live in pasting_args.R}
\usage{
pastePair(vec)
}
\arguments{
\item{vec}{}
}
\value{

}
\description{
Function that takes a pair and pastes together, with colon
Used to live in pasting_args.R
}
\keyword{internal}
