% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{pasteTwo}
\alias{pasteTwo}
\title{Function to paste every two elements in a vector together
Helper for termTraverse, regress}
\usage{
pasteTwo(vec)
}
\arguments{
\item{vec}{the vector to paste together}
}
\value{
a vector with every two elements pasted together
}
\description{
Function to paste every two elements in a vector together
Helper for termTraverse, regress
}
\keyword{internal}
