% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{polynomial}
\alias{polynomial}
\title{Create Polynomials}
\usage{
polynomial(x, degree = 2, center = mean(x, na.rm = T), version = F)
}
\arguments{
\item{x}{variable used to create the
polynomials.}

\item{degree}{the maximum degree
polynomial to be returned. Polynomials of degree <= \code{degree} will be
returned.}

\item{center}{the value to center the polynomials at.}

\item{version}{if \code{TRUE}, returns the version of the function and
nothing else.}
}
\value{
A matrix containing the linear
splines.
}
\description{
Creates polynomial variables.
}
\examples{

  # Reading in a dataset
  data(mri)
  attach(mri)
  # Create a polynomial on ldl
  polynomial(ldl, degree=3)

}
