% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{prob_KM}
\alias{prob_KM}
\title{Compute the lifetime distribution function 1 - S(t)}
\usage{
prob_KM(x, times, rightCtsCDF = TRUE)
}
\arguments{
\item{x}{an object of class \code{KM} (from a call to \code{KM})}

\item{times}{a vector of times to use in computation}

\item{rightCtsCDF}{logical, is the cdf right-continuous? (defaults to \code{TRUE})}
}
\value{
the lifetime distribution function, 1 - S(t)
}
\description{
Compute the lifetime distribution function 1 - S(t)
}
\keyword{internal}
