% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptest.R
\name{proptest}
\alias{proptest}
\alias{proptest.do}
\alias{plot.proptest}
\alias{print.proptest}
\alias{proptest.default}
\title{Test of proportions with improved layout}
\usage{
proptest(
  var1,
  var2 = NA,
  by = NA,
  exact = FALSE,
  null.hypoth = 0.5,
  alternative = "two.sided",
  conf.level = 0.95,
  more.digits = 0
)
}
\arguments{
\item{var1}{a (non-empty) numeric vector of
data values.}

\item{var2}{an optional (non-empty) numeric
vector of data.}

\item{by}{a variable of equal length to
that of \code{var1} with two outcomes. This will be used to define strata
for a prop test on \code{var1}.}

\item{exact}{If true, performs a
test of equality of proportions with Exact Binomial based confidence
intervals.}

\item{null.hypoth}{a number specifying the
null hypothesis for the mean (or difference in means if performing a
two-sample test). Defaults to zero.}

\item{alternative}{a string: one of
\code{"less"}, \code{"two.sided"}, or \code{"greater"} specifying the form
of the test. Defaults to a two-sided test.}

\item{conf.level}{confidence level of the
test. Defaults to 95/100.}

\item{more.digits}{a numeric value
specifying whether or not to display more or fewer digits in the output.
Non-integers are automatically rounded down. Any call to \code{proptest()} will
run \code{proptest.default()}, with user specified values in place of the
appropriate defaults.}

\item{\dots}{only used in the generic S3 class.}
}
\value{
Prints a summary of the data and the
corresponding proportions test. \item{Variable}{the variable
name supplied to the prop test function} 

\item{Group}{the group name: either
the variable names supplied to the function or the names of the strata if
the variable \code{by} was specified.} 
\item{Obs}{Number of observations of
each variable: includes missing values.} 
\item{Missing}{number of missing
values in each data vector.} 
\item{Mean}{the sample mean of each data
vector; also, the estimated difference in means in a two-sample test.}
\item{Std.Err.}{the estimated standard error of the mean and of the
difference in the two-sample test.} 
\item{CI}{a confidence interval for the means, and
for the difference in the two-sample test. This is at the confidence level
specified in the argument and is a two-sided confidence interval, regardless of the 
alternatuive hypothesis. If \code{prop} and/or \code{exact} are
\code{TRUE}, also returns the appropriate confidence interval for the test
of equality of proportions.}
\item{Null hypothesis}{a statement of the null
hypothesis.} 
\item{Alternative hypothesis}{a statement of the alternative
hypothesis.}
\item{t}{value of the Z-statistic, if applicable} 
\item{df}{the degrees of
freedom for the test.} 
\item{Pr}{a p-value for inference on the
corresponding hypothesis test.}
}
\description{
Produces table of relevant descriptive statistics and inference for either
one- or two-sample test of proportions. This test can be approximate or exact and has
the option to use a continuity correction.
}
\details{
Missing values must be given by \code{"NA"}s to be recognized as missing values. Any
call to \code{proptest()} is run by \code{proptest.default()}, with user specified
values in place of defaults in the appropriate places.
}
\examples{

#- Read in data set -#
psa <- read.table("http://www.emersonstatistics.com/datasets/psa.txt", header=TRUE)
attach(psa)

#- Perform t-test -#
proptest(pretxpsa, null.hypoth = 100, alternative = "greater", more.digits = 1)

#- Define new binary variable as indicator -#
#- of whether or not bss was worst possible -#
bssworst <- bss
bssworst[bss == 1] <- 0
bssworst[bss == 2] <- 0
bssworst[bss == 3] <- 1

#- Perform t-test allowing for unequal -#
#- variances between strata -#
proptest(pretxpsa, by = bssworst)

#- Perform matched t-test -#
proptest(pretxpsa, nadirpsa, matched = TRUE, conf.level = 99/100, more.digits = 1)


}
\seealso{
\code{\link[stats]{prop.test}}
}
