% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proptesti.R
\name{proptesti}
\alias{proptesti}
\alias{proptesti.do}
\alias{print.proptesti}
\alias{proptesti.default}
\title{T-test Given Descriptive Statistics with Improved Layout}
\usage{
proptesti(
  x1,
  n1,
  x2 = NA,
  n2 = NA,
  exact = FALSE,
  null.hypoth = 0.5,
  conf.level = 0.95,
  alternative = "two.sided",
  more.digits = 0
)
}
\arguments{
\item{x1}{Number of successes in first sample}

\item{n1}{Number of trials in first sample}

\item{x2}{Number of successes in second sample}

\item{n2}{Number of trials in second sample}

\item{exact}{If true, performs a
test of equality of proportions with Exact Binomial based confidence
intervals.}

\item{null.hypoth}{a number specifying the
null hypothesis for the mean (or difference in means if performing a
two-sample test). Defaults to zero.}

\item{conf.level}{confidence level of the
test. Defaults to 95/100.}

\item{alternative}{a string: one of
\code{"less"}, \code{"two.sided"}, or \code{"greater"} specifying the form
of the test. Defaults to a two-sided test.}

\item{more.digits}{a numeric value
specifying whether or not to display more or fewer digits in the output.
Non-integers are automatically rounded down.}
}
\value{
Prints a summary of the data and the
corresponding t-test. \item{Variable}{\code{x} in a
one-sample test, or \code{x} and \code{y} in a two sample test. The first
set of descriptives entered goes to \code{x}.} \item{Obs}{Number of
observations of each variable: includes missing values.} \item{Mean}{the
sample mean; also, the estimated difference in means in a two-sample test.}
\item{Std.Err.}{the estimated standard error of the mean and of the
difference in the two-sample test.} \item{Std.Dev.}{standard deviation
estimates.} \item{CI}{a confidence interval for the means, and for the
difference in the two-sample test. This is at the confidence level specified
in the argument.}
\item{Null hypothesis}{a statement of the null hypothesis.}
\item{Alternative hypothesis}{a statement of the alternative hypothesis.}
\item{t}{value of the t-statistic.} \item{df}{the degrees of freedom for the
test.} \item{Pr}{a p-value for inference on the corresponding hypothesis
test.}
}
\description{
Produces table of relevant descriptive statistics and inference for either
one- or two-sample t-test. In the two-sample case, the user can specify
whether or not equal variances should be presumed.
}
\details{
If \code{obs2}, \code{mean2}, or \code{sd2} is specified, then all three must be specified
and a two-sample t-test is run.
}
\examples{

#- Two-sample test -#
proptesti(10, 100, 15, 200, alternative = "less")

}
