% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psa}
\alias{psa}
\title{PSA dataset}
\format{
A data frame with 50 rows and 9 variables.
}
\source{
\url{http://www.emersonstatistics.com/datasets/psa.txt}
}
\usage{
psa
}
\description{
Data from a study of 50 men having hormonally treated prostate cancer. Includes information on 
PSA levels, tumor characteristics, remission status, age, and disease state.
More information, including a coding key, is available at
\url{http://www.emersonstatistics.com/datasets/PSA.doc}.
}
\keyword{datasets}
