% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{quantile_KM}
\alias{quantile_KM}
\title{Compute quantiles of the Kaplan-Meier curve}
\usage{
quantile_KM(x, probs)
}
\arguments{
\item{x}{an object of class \code{KM} (from a call to \code{KM})}

\item{probs}{a vector of proportions of interest for the quantiles}
}
\value{
quantiles of the estimated Kaplan-Meier curve
}
\description{
Compute quantiles of the Kaplan-Meier curve
}
\keyword{internal}
