% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{reFormat}
\alias{reFormat}
\title{A helper function for termTraverse}
\usage{
reFormat(current, lbls, mat, ind)
}
\arguments{
\item{current}{the current terms}

\item{lbls}{the labels of the terms}

\item{mat}{the augmented coefficients matrix}

\item{ind}{the current indices}
}
\value{
A list with an updated augmented coefficients matrix and the current indices
}
\description{
A helper function for termTraverse
}
\keyword{internal}
