% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{reFormatReg}
\alias{reFormatReg}
\title{Function to tack on args and return appropriate names for printing
Used to live in pasting_args.R}
\usage{
reFormatReg(p, h, mf)
}
\arguments{
\item{mf}{}
}
\value{

}
\description{
Function to tack on args and return appropriate names for printing
Used to live in pasting_args.R
}
\keyword{internal}
