% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_diagnostics.R
\name{rstudent.uRegress}
\alias{rstudent.uRegress}
\title{Extract Studentized residuals from \code{uRegress} objects}
\usage{
\method{rstudent}{uRegress}(model, ...)
}
\arguments{
\item{model}{an object of class \code{uRegress}, as returned by
\link[rigr]{regress}.}

\item{...}{other arguments to pass to \code{residuals.uRegress}}
}
\value{
a vector of Studentized residuals
}
\description{
Extracts Studentized residuals from \code{uRegress} objects by relying on
functionality from the \code{stats} package.
}
