% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salary}
\alias{salary}
\title{Salary dataset}
\format{
A data frame with 19792 rows and 11 variables.
}
\source{
\url{http://www.emersonstatistics.com/datasets/salary.txt}
}
\usage{
salary
}
\description{
Data from a study of 1,597 faculty members at a single US university. Includes information on 
monthly salary each year from 1976 through 1995,  as well as sex, highest degree attained, 
year of highest degree, field, year hired, rank, and administrative duties. 
More information, including a coding key, is available at
\url{http://www.emersonstatistics.com/datasets/salary.doc}.
}
\keyword{datasets}
