% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatter Plot with Lowess Curves}
\usage{
scatter(
  y,
  x,
  strata = rep(1, length(y)),
  subset = rep(TRUE, length(y)),
  reference = sort(unique(strata)),
  plotPoints = TRUE,
  plotLowess = TRUE,
  plotLSfit = FALSE,
  legend = 0.05,
  colors = c("black", "blue", "orange", "pink", "green", "red", "cornflowerblue",
    "darkolivegreen", "magenta"),
  xJitter = TRUE,
  yJitter = FALSE,
  newplot = TRUE,
  lty = 1:6,
  lwd = 1,
  log = "",
  pch = 1:25,
  ...,
  version = FALSE
)
}
\arguments{
\item{y}{\code{y} a numeric vector containing the
values to be plotted on the y-axis.}

\item{x}{\code{x} a numeric vector containing the
values to be plotted on the x-axis.}

\item{strata}{\code{strata} a vector, matrix, or list
of stratification variables. Descriptive statistics will be computed within
strata defined by each unique combination of the stratification variables,
as well as in the combined sample. If \code{strata} is supplied, all
variables must be of that same length.}

\item{subset}{\code{subset} a vector indicating a
subset to be used for all descriptive statistics. If \code{subset} is
supplied, all variables must be of that same length.}

\item{reference}{\code{reference} a list of the strata
in the order they are to be plotted.}

\item{plotPoints}{\code{plotPoints} an indicator that
points are to be plotted. A different color and point type combination will
be used for each stratum. Default value is \code{TRUE}.}

\item{plotLowess}{\code{plotLowess} an indicator that
lowess smoths are to be plotted. A different color and line type combination
will be used for each stratum. Default value is \code{TRUE}.}

\item{plotLSfit}{\code{plotLSfit} an indicator that
least squares fitted lines are to be plotted. A different color and line
type combination will be used for each stratum. Default value is
\code{TRUE}.}

\item{legend}{\code{legend} where to place the legend
on the plot}

\item{colors}{\code{colors} a vector of colors used in
plotting strata.}

\item{xJitter}{\code{xJitter} the proportion of the
minimal difference between adjacent x-values divided by 8 by which plotted
points are to be jittered in the x-dimension. A value of 0 implies no
jittering.}

\item{yJitter}{\code{yJitter} the proportion of the
minimal difference between adjacent y-values divided by 8 by which plotted
points are to be jittered in the y-dimension. A value of 0 implies no
jittering.}

\item{newplot}{\code{newplot} logical value indicating
that the graph should be plotted on a new set of axes. Default value is
\code{TRUE}.}

\item{lty, lwd, log, pch}{plotting parameters.}

\item{\dots}{\code{\dots} optional arguments for
plotting parameters (e.g. \code{xlab}, \code{ylab}, \code{main}) that will
be passed to \code{plot()}.}

\item{version}{\code{version} if \code{TRUE}, the
version of the function will be returned. No other computations will be
performed.}
}
\value{
This function produces a plot. No value is returned.
}
\description{
Produces a scatter plot of two variables with (possibly stratified)
superimposed lowess smooths and least squares fitted lines.
}
\examples{

#- Read in and attach the data -#
mri <- read.table("http://www.emersonstatistics.com/datasets/mri.txt", header=TRUE)
attach(mri)

#- Create a scatterplot with lowess curves and least squares fitted regression lines -#
scatter(atrophy, age, ylab="Atrophy", xlab="Age")

#- Create a scatterplot stratified by age on the log scale with lowess curves and -#
#- least squares fitted regression lines -#
scatter(atrophy, age, strata=male,ylab="Atrophy", xlab="Age",log="y")

}
