% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{splitOnParen}
\alias{splitOnParen}
\title{Function to split on parentheses
Helper function for termTraverse, regress
Used to live in pasting_args.R}
\usage{
splitOnParen(x)
}
\arguments{
\item{x}{the string}
}
\value{
a string with no parentheses
}
\description{
Function to split on parentheses
Helper function for termTraverse, regress
Used to live in pasting_args.R
}
\keyword{internal}
