% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip_utils.R
\name{sum_KM}
\alias{sum_KM}
\title{Compute a sum over a Kaplan-Meier curve}
\usage{
sum_KM(x, times, rightCtsCDF = TRUE)
}
\arguments{
\item{x}{an object of class \code{KM} (from a call to \code{KM})}

\item{times}{a vector of times to use in computation}

\item{rightCtsCDF}{logical, is the cdf right-continuous? (defaults to \code{TRUE})}
}
\value{
the summed survival curve, over the entered times
}
\description{
Compute a sum over a Kaplan-Meier curve
}
\keyword{internal}
