% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{termTraverse}
\alias{termTraverse}
\title{A function to traverse the termlist tree,
and create the correct augCoefficients matrix.
Helper function for regress()}
\usage{
termTraverse(termlist, mat, ind)
}
\arguments{
\item{termlist}{the list with terms}

\item{mat}{the matrix to correct}

\item{ind}{}
}
\value{
A list with the updated augmented coefficients matrix and the current indices
}
\description{
A function to traverse the termlist tree,
and create the correct augCoefficients matrix.
Helper function for regress()
}
\keyword{internal}
