% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{testList}
\alias{testList}
\title{Fully parse a formula for regress()
Takes in formula, potentially with nested F-tests}
\usage{
testList(form, modelframe, mat, data)
}
\arguments{
\item{form}{the formula to parse}

\item{modelframe}{the model frame which will become the model matrix}

\item{mat}{the matrix of logicals to run on}

\item{data}{the data frame}
}
\value{
a list of lists of tests to perform (and overall formula)
}
\description{
Fully parse a formula for regress()
Takes in formula, potentially with nested F-tests
}
\keyword{internal}
