% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{U}
\alias{U}
\title{Create a Transformed Variable}
\usage{
U(
  ...,
  type = NULL,
  subset = rep(T, length(x)),
  degree = 2,
  reference = sort(unique(x[!is.na(x)])),
  center = mean(x, na.rm = T),
  includeAll = FALSE
)
}
\arguments{
\item{...}{variable(s) used to create the
transformation.}

\item{type}{a character string describing the transformation. Partial
matching is used, so only enough of the string to make the transformation
unique is needed.}

\item{subset}{used in creating dummy variables. Only used if \code{type ==
"dummy"}.}

\item{degree}{the degree of the polynomial to be returned. Only used if
\code{type=="polynomial"}.}

\item{reference}{the reference vector for levels of the dummy variable. Only
used if \code{type=="dummy"}.}

\item{center}{the center of the returned polynomial. Only used if
\code{type=="polynomial"}.}

\item{includeAll}{a logical value to use all values even in the presense of
a subset. Only used if \code{type=="dummy"}.}
}
\value{
A matrix or vector containing the
transformations. The class of the returned value is
\code{c("transformation", y)} where \code{y} is the class of the transformed
variable (usually \code{numeric}). The type of transformation performed is
encoded as one of the attributes of the returned value, along with the
original data.
}
\description{
Creates a transformed variable using either the natural log, a dummy
transformation, linear splines, or a polynomial. Mostly for use in
regression. If a partial formula of the form \code{~var1 + var2} is entered,
returns the formula for use in regression. The partial formula can be named
by adding an equals sign before the tilde.
}
\examples{

# Reading in a dataset
mri <- read.table("http://www.emersonstatistics.com/datasets/mri.txt", header=TRUE)
attach(mri)
# Create a spline based on absolute
U(ldl, type="lspline", knots=c(70, 100, 130, 160))
U(ldl, type="ls", knots=c(70,100,130,160))

# Create a spline based on change
U(ldl, type="ls", knots=c(70, 100, 130, 160), parameterization="change")

# Create a log transformed variable
U(age, type="log")

## Create a partial formula
U(ma=~male+age)


}
\seealso{
\code{\link[uwIntroStats]{regress}}
}
