% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{uLRtest}
\alias{uLRtest}
\title{likelihood ratio test
Used to live in uLRtest.R}
\usage{
uLRtest(full, reduced)
}
\arguments{
\item{reduced}{}
}
\value{

}
\description{
likelihood ratio test
Used to live in uLRtest.R
}
\keyword{internal}
