% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uResiduals.R
\name{uResiduals}
\alias{uResiduals}
\title{Extract Residuals from uRegress objects}
\usage{
uResiduals(object, type = "", version = FALSE)
}
\arguments{
\item{object}{an object of class \code{uRegress}, as returned by
\link[uwIntroStats]{regress}.}

\item{type}{denotes the type of residuals to return. Default value is
\code{""}, which returns unstandardized residuals. \code{"standardized"},
\code{"studentized"}, and \code{"jackknife"} return the expected type of
residuals.}

\item{version}{if \code{TRUE}, the
version of the function will be returned. No other computations will be
performed.}
}
\value{
Returns the type of residuals
requested.
}
\description{
Extracts residuals (unstandardized, standardized, studentized, or jackknife)
from \code{uRegress} objects.
}
\details{
Relies on
functionality from the \code{stats} package to return residuals from the
\code{uRegress} object. \code{"studentized"} residuals are computed as
internally studentized residuals, while \code{"jackknife"} computes the
externally studentized residuals.
}
\examples{

# Reading in a dataset
data(mri)

# Create a uRegress object, regressing ldl on age
ldlReg <- regress("mean", age~ldl, data=mri)

# Get the studentized residuals
uResiduals(ldlReg, "studentized")

# Get the jackknifed residuals
uResiduals(ldlReg, "jackknife")

}
\seealso{
\code{\link[uwIntroStats]{regress}}, \code{\link[stats]{rstudent}},
\code{\link[stats]{rstandard}}
}
