% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{uWaldtest}
\alias{uWaldtest}
\title{Function to perform a Wald Test on data
Used to live in uWaldtest.R}
\usage{
uWaldtest(full, contrasts = c(0, rep(1, p - 1)), hypothesis = matrix(0, r, 1))
}
\arguments{
\item{full}{a uRegress object}

\item{hypothesis}{the null hypothesis}

\item{constrasts}{the matrix containing which objects from full to test}
}
\value{
a wald test on the given coefficients
}
\description{
Function to perform a Wald Test on data
Used to live in uWaldtest.R
}
\keyword{internal}
