\name{uwIntroStats-package}
\alias{uwIntroStats-package}
\alias{uwIntroStats}
\docType{package}
\title{
Descriptive Statistics, One Sample Inference, Regression, and Plotting in an Introductory Statistics Course
}
\description{
Developed by Scott S. Emerson, M.D., Ph.D., Andrew J. Spieker, Brian D. Williamson, Travis Y. Hee Wai, and Solomon Lim in the University of Washington Department of Biostatistics. Aims to facilitate more widespread use of R by implementing more intuitive layout and functionality for existing R functions.
}
\details{
\tabular{ll}{
Package: \tab uwIntroStats\cr
Type: \tab Package\cr
Version: \tab 0.0.3\cr
Date: \tab 2015-09-01\cr
License: \tab GPL-2\cr
}
A set of tools designed to facilitate easy adoption of R for students in introductory classes with little programming experience. Compiles output from existing routines together in an intuitive format, and adds functionality to existing functions. For instance, the regression function can perform linear models, generalized linear models, Cox models, or generalized estimating equations. The user can also specify multiple-partial F-tests to print out with the model coefficients. We also give many routines for descriptive statistics and plotting. 
}
\author{
Scott S. Emerson, M.D., Ph.D, Andrew J. Spieker, and Brian D. Williamson, Travis Y. Hee Wai, and Solomon Lim

Maintainer: Travis Y. Hee Wai <theewai@uw.edu>
}

\keyword{ package }


