% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bplot.R
\name{bplot}
\alias{bplot}
\title{Boxplot with Lowess Curves, Jittered Data,
Overlaid Mean and Standard Deviation, for an Arbitrary Number of Strata}
\usage{
bplot(
  y,
  x = rep(1, length(y)),
  data = NULL,
  strata = NULL,
  log = FALSE,
  cex = c(meancex = 1, jittercex = 0.8),
  col = c(sdcol = "dodgerblue3", jittercol = "gray30"),
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  names = NULL,
  ylim = NULL,
  legend = FALSE
)
}
\arguments{
\item{y}{dependent variable.}

\item{x}{independent variable. Must by the
same length as \code{y}.}

\item{data}{if entered, must contain both
\code{y} and \code{x}.}

\item{strata}{strata variable(s), used to
stratify the plot.}

\item{log}{a logical specifying if the data
are to be displayed on a log scale. Passed to \code{boxplot()}. Default
value is \code{FALSE}.}

\item{cex}{passed to \code{boxplot()}.}

\item{col}{passed to \code{boxplot()}.}

\item{main}{passed to \code{plot()}, the
main title of the plot.}

\item{xlab}{passed to \code{plot()}, the
x-axis label.}

\item{ylab}{passed to \code{plot()}, the
y-axis label.}

\item{names}{names (if any) of \code{x}.}

\item{ylim}{the range for plotting the
y-axis, passed to \code{plot}. Must take the form c(i, j) where the range 
i to j defines the numeric limits of the y-axis. Optionally can specificy
ylim = "full", which sets the y-axis limits to the minimum and maximum values
of the input dependent variable \code{y}.}

\item{legend}{a logical value. If \code{TRUE}, (and the means and standard
deviations have been overlaid on the graph) displays a legend next to the
first boxplot plotted denoting the max, mean+sd, mean, mean-sd, and min
values.}

\item{xjitter}{a logical specifying if
the jittered data (jittered on x-axis) are to be displayed or not. Default
is \code{TRUE}.}

\item{yjitter}{a logical specifying if
the jittered data (jittered on y-axis) are to be displayed or not. Default
is \code{TRUE}.}

\item{range}{passed to the
\code{boxplot()} function. This determines how far the plot whiskers extend
out from the box. If range is positive, the whiskers extend to the most
extreme data point which is no more than range times the interquartile range
from the box. A value of zero causes the whiskers to extend to the data
extremes.}

\item{sd}{a logical specifying if the
standard devation of \code{y} should be overlaid on the plot. Default value
is \code{TRUE}.}

\item{sdx}{a logical specifying if the
standard deviation of \code{x} should be overlaid on the plot. Default value
is \code{TRUE}.}
}
\value{
Produces a plot. No value is
returned.
}
\description{
This function adds functionality to the base R \code{\link[graphics]{boxplot}}
function. Now it is straightforward to add jittered data to the plot, and to
overlay information about the sample mean and standard deviation. The
function also supports stratification.
}
\examples{

#- Read in and attach the data -#
mri <- read.table("http://www.emersonstatistics.com/datasets/mri.txt", header=TRUE)
attach(mri)

#- Produce box plot with jittered data, sample mean, and sd -#
bplot(y=atrophy, x=male, xlab="Sex", ylab="Atrophy")

}
\seealso{
\code{\link[graphics]{boxplot}}
}
