% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_diagnostics.R
\name{cooks.distance.uRegress}
\alias{cooks.distance.uRegress}
\title{Calculate Cook's distances from \code{uRegress} objects}
\usage{
\method{cooks.distance}{uRegress}(model, ...)
}
\arguments{
\item{model}{an object of class \code{uRegress}, as returned by
\link[rigr]{regress}.}

\item{...}{other arguments to pass to \code{stats::cooks.distance}}
}
\value{
a vector of Cook's distances
}
\description{
Extracts Cook's distances from \code{uRegress} objects by relying on
functionality from the \code{stats} package.
}
