% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fev}
\alias{fev}
\title{FEV dataset}
\format{
A data frame with 654 rows and 7 variables:
\describe{

\item{seqnbr}{case number (the numbers 1 to 654)}

\item{subjid}{subject identification number (unique for each different child)}

\item{age}{subject age at time of measurement (years)}

\item{fev}{measured forced exhalation volume (liters per second)}

\item{height}{subject height at time of measurement (inches)}

\item{sex}{subject sex}

\item{smoke}{smoking habits ("yes" or "no")}
}
}
\source{
\url{http://www.emersonstatistics.com/datasets/fev.txt}
}
\usage{
fev
}
\description{
Data from a study of 654 children on the relationship between smoking status and 
lung function (measured by FEV). Each row corresponds to a single clinic visit and contains
information on age, height, sex, FEV, and smoking status.
More information, including a coding key, is available at
\url{http://www.emersonstatistics.com/datasets/fev.doc}.
}
\keyword{datasets}
