% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_diagnostics.R
\name{hatvalues.uRegress}
\alias{hatvalues.uRegress}
\title{Calculate the hat-values (leverages) from \code{uRegress} objects}
\usage{
\method{hatvalues}{uRegress}(model, ...)
}
\arguments{
\item{model}{an object of class \code{uRegress}, as returned by
\link[rigr]{regress}.}

\item{...}{other arguments to pass to \code{stats::hatvalues}}
}
\value{
a vector of hat-values (leverages)
}
\description{
Extracts hat-values (leverages) from \code{uRegress} objects by relying on
functionality from the \code{stats} package.
}
