% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psa}
\alias{psa}
\title{PSA dataset}
\format{
A data frame with 50 rows and 9 variables:
\describe{

\item{ptid}{patient identifier}

\item{nadirpsa}{lowest PSA value attained post therapy (ng/ml)}

\item{pretxpsa}{PSA value prior to therapy (ng/ml)}

\item{ps}{performance status (0= worst, 100= best)}

\item{bss}{bone scan score (1= least disease, 3= most)}

\item{grade}{tumor grade (1= least aggressive, 3= most)}

\item{age}{patient's age (years)}

\item{obstime}{time observed in remission (months)}

\item{inrem}{Indicator whether patient still in remission at last follow-up (yes or no)}

}
}
\source{
\url{https://rct-design.com/TeachingMaterials/Datasets/psa.txt}
}
\usage{
psa
}
\description{
Data from a study of 50 men having hormonally treated prostate cancer. Includes information on 
PSA levels, tumor characteristics, remission status, age, and disease state.
More information, including a coding key, is available at
\url{https://rct-design.com/TeachingMaterials/Datasets/psa.txt}.
}
\keyword{datasets}
