% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salary}
\alias{salary}
\title{Salary dataset}
\format{
A data frame with 19792 rows and 11 variables:
\describe{

\item{case}{case number}

\item{id}{identification number for the faculty member}

\item{sex}{M (male) or F (female)}

\item{deg}{highest degree attained: PhD, Prof (professional degree, eg, medicine or law), or Other (Master's or Bachelor's degree)}

\item{yrdeg}{year highest degree attained}

\item{field}{Arts (Arts and Humanities), Prof (professional school, e.g., Business, Law, Engineering or Public Affairs), or Other}

\item{startyr}{year in which the faculty member was hired (2 digits)}

\item{year}{year (2 digits)}

\item{rank}{rank of the faculty member in this year: Assist (Assistant), Assoc (Associate), or Full (Full)}

\item{admin}{indicator of whether the faculty member had administrative duties (eg, department chair) in this year: 1 (yes), or 0 (no)}

\item{salary}{monthly salary of the faculty member in this year in dollars}

}
}
\source{
\url{http://www.emersonstatistics.com/datasets/salary.txt}
}
\usage{
salary
}
\description{
Data from a study of 1,597 faculty members at a single US university. Includes information on 
monthly salary each year from 1976 through 1995,  as well as sex, highest degree attained, 
year of highest degree, field, year hired, rank, and administrative duties. 
More information, including a coding key, is available at
\url{http://www.emersonstatistics.com/datasets/salary.doc}.
}
\keyword{datasets}
