% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_utils.R
\name{U}
\alias{U}
\title{Create a Partial Formula}
\usage{
U(...)
}
\arguments{
\item{...}{partial formula of the form \code{~var1 + var2}.}
}
\value{
A partial formula  (potentially named) for use in \code{\link[rigr]{regress}}.
}
\description{
Creates a partial formula of the form \code{~var1 + var2}. The partial formula can be named
by adding an equals sign before the tilde.
}
\examples{

# Reading in a dataset
data(mri)

# Create a named partial formula
U(ma=~male+age)

# Create an unnamed partial formula

U(~male+age)

}
\seealso{
\code{\link[rigr]{regress}}
}
